modularized back end and created auth routes
